fixed team editor name bugs, added debug logger to PlayerDisplay for advancement progress, fix small card task locations

+++ BingoReloaded/src/main/java/io/github/steaf23/bingoreloaded/data/BingoCardData.java
@@ -86,8 +86,8 @@ public class BingoCardData
 
     public void setList(String cardName, String listName, int max, int min)
     {
-        data.setInt(cardName + "." + listName + ".max", max);
-        data.setInt(cardName + "." + listName + ".min", min);
+        data.setByte(cardName + "." + listName + ".max", (byte) max);
+        data.setByte(cardName + "." + listName + ".min", (byte) min);
         data.saveChanges();
     }
